Refactored integ-test classes.
